Completelly got rid of sse utils header.
